# Optimize DOM traversal