… oracles/convertTeaspoon.html
… oracles/default.html
… oracles/degWithoutFahrenheit.html
… oracles/includeImproperSymbols.html
… oracles/includeQuotes+includeImproperSymbols.html
… oracles/includeQuotes.html
… oracles/isUK.html
… oracles/matchIn.html
… oracles/useBold.html
… oracles/useBrackets+useBold.html
… oracles/useBrackets.html
… oracles/useComma.html
… oracles/useGiga.html
… oracles/useKelvin.html
… oracles/useMM.html
… oracles/useRounding.html
… oracles/useSpaces.html