Hi again @m1L!

Following my last comment on #40, I took a look at the DOM traversal logic. It looked like calling `node.nextSibling` was costing a significant amount of time. With some Stack Overflow sleuthing and random keyword googling, I learned that there is actually an API for this: [`TreeWalker`](https://developer.mozilla.org/en-US/docs/Web/API/TreeWalker)! By using `NodeFilter.SHOW_TEXT`, we can quickly select the text nodes.

The issue is that this breaks the filtering on ancestor nodes. We could use the `filter` argument of `TreeWalker`, but this is slow. Or call `hasParentEditableNode`, but this is even slower. So I have used another approach: when starting a traversal, I first list all the nodes that should be excluded, using CSS selectors; then, during the traversal, I check whether the text node's parent is in the exclusion list. It seems to work pretty well.

After doing that, there was a significant amount of time spent on `node.parentNode`. This was needed, since we cannot list text nodes directly from the CSS selectors, so we need to check whether each text node's parent is in the exclusion list. However, I also learned about [`XPathEvaluator`](https://developer.mozilla.org/en-US/docs/Web/API/XPathEvaluator), which is similar to CSS selectors, but allow us to select text nodes directly. I was able to replace CSS selectors and get rid of `node.parentNode`.

However, `XPathEvaluator` is slower than CSS selectors, so both approaches are roughly as efficient. I have kept both commits in the history, so that we can go back to CSS selectors if it turns out they are better.

---

I have also added some (manual) tests to `index.html` for this. The mutation observer should also handle the insertion of text nodes, not just of element nodes.

---

With this, the [Web version of _From the Earth to the Moon_](https://www.gutenberg.org/cache/epub/83/pg83-images.html) goes from taking 174 ms to only 106 ms. The [text version](https://www.gutenberg.org/cache/epub/83/pg83.txt) goes from 298 ms to 279 ms, but I think it might just be noise.

Before:

![image](https://github.com/m1l/Everything-Metric-Firefox/assets/8493765/e7eb4885-4d66-4a16-8965-bbdb2ee18348)

After:

![image](https://github.com/m1l/Everything-Metric-Firefox/assets/8493765/dc09f064-721d-4a5e-adbf-c4d936a24e1d)

Most of the remaining time is spent in `exec(…)` and in the calls to `toLocaleString(…)`. For the first one, merging all the `replace*` functions into one, and using a single pass with a single regular expression should gain us some performance. For the second one, the main point is related to #41.

![image](https://github.com/m1l/Everything-Metric-Firefox/assets/8493765/13e6bad3-a599-4372-8ed4-efa8ce5fc15a)