@m1l 2023-10-21
@qsantos  Choose the one that you think is the best as you are much further ahead with this. It's good that there are options. 

I don't know if this would make difference by using these approaches, but one thing I failed in my approach is inserting nodes while traversing and updating units. I think that an ultimate goal (which is also suggested in one of the opened issues) is to be able to have an option to replace all imperial units completely (once we are certain it works well), and then have original unit available in a tooltip on hover of the converted unit. 

Some will still prefer seeing both and will probably trust it more with all the mess in how imperial units are used (especially when shopping for something that a critical fit, to be able to verify using both units), but most will probably opt into just seeing international units.