@qsantos 2023-10-22
At a later point, I will investigate simply setting `innerHTML` to insert new element nodes. Surprisingly, it can actually be faster to generate HTML and let the browser handle it than use many calls to `document.createElement` and `node.appendChild`. Of course, it might depend on the situation and may not even be true anymore, so I will have to do some benchmarks. Yay! :tada: